backend-improve/requirements.txt
# Core dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
python-dotenv>=1.0.0
pydantic>=2.3.0

# Gemini AI and ADK
google-generativeai>=0.3.1
google-cloud-aiplatform>=1.36.3

# For ADK mode
google-adk>=0.1.0

# For Tavily Search
langchain>=0.1.0
langchain-community>=0.1.0
tavily-python>=0.2.8

# Utility
requests>=2.31.0
aiohttp>=3.8.5
